move mssql init schema logic to ee
